Revert "Merge pull request #19664 from vespa-engine/jonmv/streaming-doc-v1-visit"
This reverts commit b5d4b42caabf2c41fcbd8b21814819aae77dc7aa, reversing
changes made to 9abe019606f2367b05e4e13d796de65dddf7c449.
